+++ Src/Web/Web.Shared.Net/RequestTrackingTelemetryModule.cs
@@ -5,6 +5,7 @@
     using System.Globalization;
     using System.Web;
 
+    using Microsoft.ApplicationInsights.Common;
     using Microsoft.ApplicationInsights.Extensibility;
     using Microsoft.ApplicationInsights.Extensibility.Implementation;
     using Microsoft.ApplicationInsights.Web.Implementation;
@@ -16,6 +17,7 @@
     {
         private readonly IList<string> handlersToFilter = new List<string>();
         private TelemetryClient telemetryClient;
+        private bool correlationHeadersEnabled = true;
 
         /// <summary>
         /// Gets the list of handler types for which requests telemetry will not be collected
@@ -28,6 +30,22 @@
                 return this.handlersToFilter;
             }
         }
+
+        /// <summary>
+        /// Gets or sets a value indicating whether the component correlation headers would be set on http responses.
+        /// </summary>
+        public bool SetComponentCorrelationHttpHeaders
+        {
+            get
+            {
+                return this.correlationHeadersEnabled;
+            }
+
+            set
+            {
+                this.correlationHeadersEnabled = value;
+            }
+        } 
 
         /// <summary>
         /// Implements on begin callback of http module.
@@ -84,7 +102,44 @@
                 requestTelemetry.Url = context.Request.UnvalidatedGetUrl();
             }
 
+            if (context.Request.Headers != null)
+            {
+                // If the source header is present on the incoming request, use that to populate the source field.
+                string sourceIkey = context.Request.Headers[RequestResponseHeaders.SourceInstrumentationKeyHeader];
+
+                if (!string.IsNullOrEmpty(sourceIkey))
+                {
+                    requestTelemetry.Source = sourceIkey;
+                }
+            }
+
             this.telemetryClient.TrackRequest(requestTelemetry);
+        }
+
+        /// <summary>
+        /// Adds target response header response object.
+        /// </summary>
+        public void AddTargetHashForResponseHeader(HttpContext context)
+        {
+            if (this.telemetryClient == null)
+            {
+                throw new InvalidOperationException();
+            }
+
+            var requestTelemetry = context.GetRequestTelemetry();
+
+            if (string.IsNullOrEmpty(requestTelemetry.Context.InstrumentationKey))
+            {
+                // Instrumentation key is probably empty, because the context has not yet had a chance to associate the requestTelemetry to the telemetry client yet.
+                // and get they instrumentation key from all possible sources in the process. Let's do that now.
+                this.telemetryClient.Initialize(requestTelemetry);
+            }
+
+            if (!string.IsNullOrEmpty(requestTelemetry.Context.InstrumentationKey)
+                && context.Response.Headers[RequestResponseHeaders.TargetInstrumentationKeyHeader] == null)
+            {
+                context.Response.Headers[RequestResponseHeaders.TargetInstrumentationKeyHeader] = InstrumentationKeyHashLookupHelper.GetInstrumentationKeyHash(requestTelemetry.Context.InstrumentationKey);
+            }
         }
 
         /// <summary>