+++ Src/WindowsServer/WindowsServer.Nuget.Tests/TelemetryModulesTests.cs
@@ -65,6 +65,36 @@
         }
 
         [TestMethod]
+        [Ignore]
+        public void InstallAddsFirstChanceExceptionTelemetryModule()
+        {
+            string emptyConfig = ConfigurationHelpers.GetEmptyConfig();
+            XDocument configAfterTransform = ConfigurationHelpers.InstallTransform(emptyConfig);
+
+            Type typeToFind = typeof(FirstChanceExceptionStatisticsTelemetryModule);
+
+            var node = ConfigurationHelpers.GetTelemetryModules(configAfterTransform)
+                .Descendants()
+                .FirstOrDefault(element => element.Attribute("Type").Value == ConfigurationHelpers.GetPartialTypeName(typeToFind));
+
+            Assert.IsNotNull(node);
+        }
+
+        [TestMethod]
+        public void InstallDoesNotInstallExtraModules()
+        {
+            string emptyConfig = ConfigurationHelpers.GetEmptyConfig();
+            XDocument configAfterTransform = ConfigurationHelpers.InstallTransform(emptyConfig);
+
+            Type typeToFind = typeof(FirstChanceExceptionStatisticsTelemetryModule);
+
+            ////Assert.AreEqual(4, ConfigurationHelpers.GetTelemetryModules(configAfterTransform).Descendants().Count());
+
+            // change to 4 after re-enabling first chance exceptions
+            Assert.AreEqual(3, ConfigurationHelpers.GetTelemetryModules(configAfterTransform).Descendants().Count());
+        }
+
+        [TestMethod]
         public void InstallAddsUnobservedExceptionTelemetryModule()
         {
             string emptyConfig = ConfigurationHelpers.GetEmptyConfig();